mac-capture: Fix display capture cursor bug
You cannot use @true or @false with the dictionary, after some testing,
it needs to point to the id of either kCFBooleanTrue or kCFBooleanFalse.

This solves the bug where 10.10+ would not render the cursor, because it
was always using the internal default value, and was ignoring the value
we passed to it.